Numerical Differential Equations | Simulating a standing wave
Our solver2 algorithm involves taking a given number of displacement steps to take - N, given number of time steps - M, total length of system - L and total time of wave motion - T.

Following that, since we wish to see how the approximated solution looks like in intervals, **we set a variable named 'check' to be the number of timesteps the algorithm calculates, before stopping to plot the solution periodically**. If the given timestep is an integer multiple of the given 'check' variable, the algorithm plots the solution at that timestep, and does this again once it reaches the next integer multiple.

Following that, since there are undoubtedly instances when we wish to only see plots - without any outputted results at the end, others when we only wish **to see the final calculated approximation, we set a 'req' variable**. This variable, is by default set to 0, and at that setting, will not provide any output of the approximated solution at the final time step. If set to any value other than 0, it produces that output, as well as the x axis vector to plot against - X, the list of the timesteps at which the solution was stored into the list 'savedata' - 'dataindex'. Conceivably, one could choose to not have the solver2 algorithm automatically plot the solution at every single timestep, and instead, look through the saved data themselves.